Added ability to get hint information from TextViews.
Modified ShadowTextView.java to inflate hint information from layout.
Modified TextViewTest.java to test that hint information is being inflated.
Added text_views_hints.xml for a test layout.
Modified R.java to see the additions in text_views_hints.xml. I don't know if this is the right way to do this, but it was the only way that worked for me.