Make android.os.Vibrator work with JellyBean
Summary: android.os.Vibrator became an abstract class in Android 4.1.
This copies the dummy NullVibrator implementation from the 4.1 sdk
source and instructs robolectric to load that class if it sees that
android.os.Vibrator is abstract.

Test Plan: Linked against the r16 (jellybean) android.jar and ran unit
tests which relied on Vibrator.