Added ability to handle intercepted calls to LinkedHashMap.eldest().
Setup can return handler objects for class name / method name combinations.
ShadowWrangler delegates to these handlers when nasty method calls are intercepted.
LinkedHashMap.eldest() (which is a VERY nasty method that is not defined to exist in the JDK (shame on you Google)) is associated with a handler in Setup.