Issue #647: Correctly match resource directories.
This updates the PackageResourceLoader so that it correctly excludes
directories that aren't the standard resource directories.  This is necessary
in case a path is configured like:

android-menudrawer/menudrawer/res/values/attrs.xml

It would incorrectly in the past try to include attrs.xml in as part of the
MenuLoader parsing becuase it contained "menu" in the path name.  This patch
appends the "res/" to the folderBaseName passed to the resource loader to
limit the scope further.