Fixes NullPointerException when using @Config manifest NONE
When testing with ContentResolver if @Config(manifest=Config.NONE) is
used then getProvider() method will get into a NullPointerException when
attempting to iterate through the list of the AndroidManifest's
ContentProviders.

Wrapped the statement in a null check to fix the issue.
Added a new test to cover this use case.