Changed Converter to accept values directly for enums and flags.
This comes up as an issue when an xml attribute refers directly to
the ordinal of an enum rather than the name of the enum. For example,
this is a problem with preference_list_fragment.xml, where Android
refers explicitly to the int 2 rather than 'outsideOverlay'. Since
the Android docs expressly indicate the enum ordinals pairing with
the name, this seems like reasonable behavior to expect.

Fixes #992. Closes #1092.