Make fest-util a test only dependency.
The fest-util that is used by Robolectric has version skew with both fest-reflect
(as seen by the exclusion in Robolectric's pom.xml) and also fest-assert.

Unfortunately many projects here use a version of fest assert that is runtime incompatible
with the version of fest-util that ships with Robolectric and this is blocking our
company wide upgrade to the latest version of Robolectric.

Until there is a public version of fest (assert,reflect,util) that is free of version skew
I reccomend against including fest-util as a production dependency of Robolectric.