Implement `ContentResolver.isSyncActive(Account,String)`
Currently this falls to the shadowed implementation, which fails
with `NullPointerException` because `getContentService()` returns
`null`.

Since the Shadow is already keeping track of requested syncs, we
can use that to determine if a sync is "active" or not.

One caveat: after the initial `requestSync()` call, all future
calls to `isSyncActive()` will return true, because there is no
way to stop (or simulate stopping) a "sync request" in the shadow.