Fix ShadowIntent#putExtra(String, byte[])'s return type.
ShadowIntent's putExtra method for byte arrays was marked as a void
method, when in Android this method actually returns an Intent (the same
Intent, so that these methods are chainable). I've updated the return
type and value to match Android, and added a test to guarantee that all
putExtra methods are chainable.