Support custom item states in android selectors.
This was failing because custom namespaced attributes are not correctly resolved, since the Node.getName()
returns the fully qualified name including the namespace which cannot be resolved in Robolectric's ResourceIndex.

We actually need to check here if the attribute's namespace is the default namespace and if so then we can just
return the local name.

This PR is also related to https://github.com/robolectric/robolectric/pull/1272 which will
replace any res-auto custom namespaces with the default package.