Delete unnecessary ShadowKeyEvent.
This implementation got in the way of KeyEvent dispatch logic (which
uses the mAction and mKeyCode fields, which weren't being set by the
shadowed constructor). The ShadowKeyEvent wasn't exposing any behavior
that isn't already implemented by KeyEvent and/or ShadowInputEvent, so
deleting it and using the real constructors seems to be the sensible
option.

This keeps the test around, but renames it to InputEventTest, since
that's where the behavior lies that it was actually testing.