Fix for https://github.com/robolectric/robolectric/issues/1485
Lean on real android for creation of application data dirs - The Android environment provides and creates a "data dir" for the installed application through the ApplicationInfo. Change Robolectric infrastructure to create and assign this value and remove custom shadow code that manages this.

Remove ShadowContextWrapper.getApplicationInfo() we should delegate to real android for this (which provides us with the ApplicationInfo we create in the DefaultPackageManager) instead of creating a new instance with limited content.

ParallelUniverse ensures the application's data dir is removed at the end of the test.

Fix up tests that relied on ShadowContextWrapper.getApplicationInfo(), now that we rely on real Android these components must be attached to a base activity.