Fix bug introduced in https://github.com/robolectric/robolectric/commit/df98189
While it is true that NfcAdapter.getDefaultAdapter() queries the system service, this returns the NfcManager which calls back into NfcAdapter.getNfcAdapter() this makes several binder calls and basically NPEs, so needs shadowing to return a valid instance. Also, NfcManager needs calling with a context as its constructor parameter otherwise it is not initialized correctly.

Lastly, add support for NfcAdapter.setNdefPushMessageCallback() and a non-android accessor so the callback can be executed in tests.