Initial version of a test runner that will run Robolectric tests on multiple API versions.
Based on a ParameterizedTestRunner, these test runners shard at the test runner level, so
one TestRunner is created for each API level. We will be able to apply class level @Config
emulateSdk filtering here, but not method level, e.g: if a method has emulateSdk on it then
it will still exist for each TestRunner, but maybe it can be considered ignored?

The alternaive it to do it at the method level and override runChild() and for each API
version then call RobolectricTestRunner.methodBlock()

Comments appreciated.