Fix leaky test runners.
Both classes are overriding key behaviour in the parent test runner that  nulls out the test lifecycle preventing it being eligible for gc. This is particularly nasty for the Robolectric suite of tests as the TestLifeCycle holds a reference to its class which in turn holds a reference to its InstrumentingClassloader and all its loaded classes which eats up a lot of permgen.