Fix tests and shadows depending on errant equals() behavior in Intent.
This also fixes a test that incorrectly verified that an Intent with no
categories set would return an empty list - the real documented
behavior of that method, unfortunately, is to return null.

https://developer.android.com/reference/android/content/Intent.html#getC
ategories()

Most of the code here was modified to use filterEquals - for launching
and tracking services and activities, this is what we really want.
There’s still some value in a real equals comparator for intents though
- some day I may get to it.