Remove wrongly added Test7 from @Config tests
Commit 6b40d430fe0e48aa0ccaf9f95206ac70bdf6732c fixed incorrect
overriding behaviour of the constants config property, and also modified
the various config overriding tests to include this property.
Unfortunately, for the tests for loading the configurations from the
properties file, instead of just adding and verifying the constant
property in the mock file, it replaced the non-annotated mock class
(Test2) that was used in the test with a new mock class that had the
constants property defined via annotation (Test7). This basically
invalidated the test, and caused the fixed issue to no longer be covered
by it. This commit fixes that mistake by removing Test7, and reverting
it's referencing tests to use Test2, and also adding the constants
property in the mock property file test.