Test ParameterizedRobolectricTestRunner with Android class parameters
There is a known issue that the runner does not work when the test parameters are objects of Android classes, and are tests against objects which are created during the tests. This is because the class loader of the original parameters, and the objects created during the tests are different.