Fix view animation scheduling mechanism
Commit 7b1c2d95d341eac22e41e318ddce6ff373d9a04a shadowed the
View#setAnimation() method in order to provide a simulated functionality
despite the drawing mechanism being broken. However, it provided the
start time as the delay interval, and did not account for the start
offset. Moreover, it performed the animation in a loop without
delegating to the Choreographer or Scheduler.

This commit fixes these issues by creating a simulator class that
schedules it at the correct start time, and also performs the animation
via the scheduler. The animation time incrementation is now performed
internally instead of by the ShadowSystemClock tracker, since there is
no way to set it to the correct animation starting time.

The removeCallbacks() method in Choreographer has also been shadowed to
work with the shadowed postCallbackDelayed() method implementation, so
that the animation callbacks can be removed when a new animation is set.