Update ShadowAlarmManager.java.vm
Shadowing of this class is done not really correctly. I was using it for testing the alarms. I had to set up alarms at different time that will start the same BroadcastReceiver. The alarms fire properly one after each other but the ShadowAlarmManager cancels the ones that was scheduled before and shows that only one alarm was scheduled. It happens because the Intent.filterEquals() returns true in this case but it doesn't check if the time is the same.