Fix install-dependencies.rb maps version
The correct version string should look like "18_r4", not just "4".

This bug was introduced in bf4f3aa (issue #1968 / PR #2092) by accident.

Note that this also intentionally uses #{revision} instead of
the shell version did. (This resulted in the previous version actually
trying to install to filenames with "4?" instead of "4".)