Add a shadow for RenderNodeAnimator.
RenderNodeAnimator is used by RevealAnimator, which is the Animator
that ViewAnimationUtils creates for circular reveals (21+). Since it
uses native code for actually rendering and timing the animation, it
requires a shadow if we want these animations to actually run/finish.

The shadow implementation here utilizes the Choreographer to time
itself, so it has the same characteristics of other Animators
(ObjectAnimator, ValueAnimator) with regard to testing. The 21 version
of this class had some bugs, unfortunately, but this Shadow reproduces
these faithfully so that tests are accurate.