Implement Context.startIntentSender()
This was required because after the removal of ShadowRemoteViews the real RemoteViews uses Context.startIntentSender() rather than Context.startActivity() etc.

Set the application in the LoadedApk so that it is available for ContextImpl.getApplicationContext()

Implement ShadowContextImpl.{startActivity,startService,sendBroadcast} and delegate through to the application. We should actually move the functionality away from ShadowApplication and to ShadowContextImpl as this is where the actual code invoked for Application is as it doesn't have its own implementation and goes via its parents implementation (ContextWrapper) to the ContextImpl.