Fix a ConcurrentModificationException
If resources have not been initialized, and two threads try to access a
resource simultaneously, they both try to initialize the same
ResourceLoader.  This can lead to two issues:

1. One thread modifies a list will the other is iterating over it,
throwing a ConcurrentModificationException, as in #1564
2. One thread finishes, calling ResBundle#makeImmutable(), and then the
other thread tries to call ResMap#merge(), throwing an
IllegalStateException, as in #2158

By simply marking XResourceLoader#initialize() as synchronized, we
prevent both these issues.