Service controller API Improvements (#2406)
* ServiceController should automatically attach the base context.

* ActivityController should automatically attach the base context.

Add new factory methods that take an Intent and set in the ComponentController at construction time. This avoids the dificult to diagnose problem of withIntent() not working if it is called after ActivityController.attach(). @Deprecate ActivityController.attach() to migrate away from this broken API. Test writers should not need to concern themselves with attaching an activity as its a part of the lifecycle that is transparent to app developers. Provide a hacky way of setting the intent after onAttach() has been called to help with the migration. This will be removed when we remove onAttach() and withIntent() from the public API.

Add new convenience factory methods for FragmentController.