Set Build.Version.RELEASE and Build.Version.INCREMENTAL
This patch sets RELEASE and INCREMENTAL to non-null values when
running Robolectric. I chose two somewhat arbitrary values; in
particular, I chose "2.2" for the version string to keep it in
sync with SDK_INT.

It would be cool if the values depended on the selected SDK
though. I guess it's a matter of making the shadow API-dependent?