Synchronize the plan cache in ShadowWrangler (#2458)
The plan cache in ShadowWrangler is used concurrently when running
tests under gradle. Adding and removing plans from the cache
occasionally incurs in a race condition that causes test flakiness.

Twitter uses a custom version of OpenJdk that makes this issue
particularly noticeable, since it nulls out an entry's links when
it gets removed. In this case, the race condition causes and NPE.
In the general case, the map uses entries that are already removed.

I fixed by making the map synchronized and replacing containsKey/get
with a single call to get and a null check. I considered making the
method synchronized, but it felt wasteful to run plan calculation
within a synchronized block. The only downside of this is that the
occasional plan may get calculated twice.