When loading an XML resource, if the resource is a framework resource then it should use the system AssetManager that contains the resource details from the runtime framework. On a real device the resource table for system resources is encoded in the arsc file of framework-res.apk. If the resource comes from the app then it requires the apps resource details which are stored in the arsc file in the apps apk.
The apps resource arsc file is built at compile time and is based on the resources from the app and its dependency libraries and also from the version of Android SDK it was compiled against. This means when running on an older device it is able to understand public Android framework resources from the later version it was compiled against.