Remove unnessesary usage of converter from RoboAttributeSet. (#2505)
Add an undefined attribute used in RoboAttributeSetTest to the resources so that we can remove the defaulting code used to paper over that condition. i.e: If you use RoboAttributeSet then it must refer to actual resources. This allows us to use an alternative Converter.convertAndFill() method and make one private.