Tweak ResBundle matching algorithm to handle chaining
ResBundle#pick never returns null. If nothing
matches, it just returns the first thing in
the list of choices.

This causes a problem if resources are separated
into two chained bundles (the case with anydpi
vector xml and image drawables). It means that
the second bundle never gets a chance to be
considered.