Parse manifest file immediately to stabilize hashcode
The hashCode function for the AndroidManifest class depends on the
package name. Initially, if an AndroidManifest object is constructed
without providing a package name, the initial package name is null.
After `parseAndroidManifest` is called, the packageName is extracted
from the manifest file, and `hashCode` will return a different value.

In certain situations, this broke resource caching in
RobolectricTestRunner. The cache is essentially a HashMap with an
AndroidManifest object as part of the key. Because the value of
`hashCode` changes, the lookups would not work properly.

To alleviate the issue, parse the manifest file immediately after
construction. This ensures that all components of the hashCode method
are provided at construction.

Fixes #2623