Adjust install directories for robolectric-shadows projects
Previously, when performing a `./gradlew install`, all projects in the
robolectric-shadows directory were installed to the maven path with the
parent directory `org/robolectric/robolectric-shadows/`. This is not
consistent with official artifact ids of these projects, and caused
build failures when trying to build the 3.2-SNAPSHOT version of these
projects. Instead, install them to the local path with the parent
directory `org/robolectric`.

After this change, it's possible to include 3.2-SNAPSHOT projects such
as shadows-support-v4 using:

    testCompile 'org.robolectric:shadows-support-v4:3.2-SNAPSHOT'

Fixes #2696