Add ShadowIcon
There are some hidden public methods on the Icon class that are useful
for test-related assertions.

For example, if a test displays a notification, it may be useful to
examine the icons in the notification or the icons that belong to
actions in the notification.

Currently tests for this class require Android L+ due to the reference
to Drawable.DEFAULT_TINT_MODE, which is only available in Android L and
above.