Add support for @Resetter in user-defined shadows
Shadows processed using the Robolectric annotation processor have
support for the @Resetter annotation. This is a convenient way for a
shadow class to reset static state before each test is run.

Previously, this was only available for shadows processed using the
annotation processor. Add support for @Resetter annotations in
user-defined shadows.

Issue: #1072