Call standard View.getGlobalVisibleRect() is the view is laid out
getGlobalVisibleRect() behaves correctly if the view is part of a view
hierarchy and has been laid out. The current implementation of this
method in ShadowView makes the assumption that the view is not laid out,
and avoids using the standard implementation, which causes bugs in some
tests.

This patch calls getGlobalVisibleRect() in the SDK whenever the view has
been laid out. I left the original code in place so that my change won't
break existing tests that may rely on this behavior - I think, however,
that tests should ideally lay out views instead of relying on non standard
behavior.