Stop scheduling background tasks relative to foreground scheduler's current time
When calling Handler#post(Runnable) and variants, the scheduled time on
the background looper is relative to the foreground scheduler's current
time. This is because Handler#post(Runnable) ultimately make its way
down to Handler#sendMessageDelayed(Message, long), which invokes
SystemClock.uptimeMillis(), which returns the time relative to the
foreground scheduler's current time.

One side effect is that, in some cases, ShadowLooper#runToEndOfTasks may
be a no-op. This happens when scheduling a task on a background handler
when its scheduler has a greater current time than the foreground
scheduler.

Add an implementation of Handler#sendMessageDelayed, which schedules the
task relative to the Handler's current time.

At first I attempted to avoid scheduling tasks that occur in the past
(either in ShadowMessageQueue#enqueueMessage or in
Scheduler#queueRunnableAndSort), but that resulted in failing tests.

One side effect of this is that it un-deprecates ShadowHandler.

Fixes #2957