Make ShadowPackageManager not abstract.
This allows teams to subclass PackageManager without requiring to have a
corresponding shadow.

The methods it needs to implement come from the RobolectricPackageManager
interface which will be removed in 3.4 These are all methods that are
available on the framework PackageManager which should be used in
preference to the shadow anyway.