Add support for `targetSdkVersion="O"`
According to the Android migration docs, during the O preview, apps are
encouraged to use `targetSdkVersion="O"` instead of a numeric SDK
version. If this is added to the manifest, Robolectric is unable to
parse the targetSdkVersion, and the min sdk version is used instead.
This mismatch may cause problems because the runtime SDK believes the
app is targeting an older SDK version than it actually is.

Once Android O officially launches, this can be removed.

See https://developer.android.com/preview/migration.html for details.