Use real Android's SharedPreferences implementation.
RoboSharedPreferences was an in-memory map, real Android uses a file.
This caused a bug where Context.moveSharedPreferencesFrom() did not work
which is fixed when using real Android. Added a test to assert this.

Also exposed a bug where RoboSharedPreferences.Editor.commit() would
apply clear() in the order it was called, wiping previous values in the
editor. In real Android clear() is always applied first preserving any
values in the current Editor.

On Android O file writes are added to a separate QueuedWork thread for
some writes so we must wait for this to finish between tests. This is
something that should probably be done in the Robolectric environment.