Reset shared preferences static cache. (#3055)
* Reset shared preferences static cache.

Prevents test pollution. On some platform versions this is initialized
to a HashMap, on later versions it is lazy initialized to an ArrayMap,
the field name has also changed across API versions.

* Shadow IoUtils.readFileAsString()

On API < 19 this class references java.lang.UnsafeByteSequence so have
Robolectric re-write this as Object since it is only present in the
Dalvik VM. Shadow for all API levels however as post API 19 it still
uses a lot of libio stuff.