Reset all Loopers' schedulers in ShadowLooper's resetter
Previously, it was possible for tasks posted to a static Handler to leak
across tests. This was because background Loopers usually retained their
previous Scheduler state and Runnables (except after the first
invocation of the resetter when Loopers are quit).

Instead, call Scheduler#reset for each ShadowLooper. This clears any
posted tasks and clears the state of the Scheduler. Before this was only
done for the main Looper.

Fixes #2962