Invoke zero-arg constructor when building an IntentService
Previously, Robolectric#buildServiceIntent invoked the one-arg
constructor of the IntentService subclass. If the IntentService subclass
did not have a one-arg constructor, a NoSuchMethodException was thrown.

An Android IntentService subclass is not really required to have a
one-arg constructor because the IntentService class provides a default
one that is usually called from the zero-arg constructor.

Instead, invoke the zero-arg constructor, which is required for all
Android Services. In most cases, the zero-arg constructor of the
subclass invokes the one-arg constructor of IntentService. This is more
consistent with the expectation in Android.

Fixes #3080