Add looseSignatures=true to ShadowNotificationManager
ShadowNotificationManager was importing NotificationChannel and
NotificationChannelGroup. These classes are only available starting in
Android O. If an app that targets SDK 25 below attempted to use
Robolectric 3.4-rc1's ShadowNotificationManager, a
java.lang.NoClassDefFoundError would be thrown.

To fix this, add looseSignatures = true to ShadowNotificationManager,
and convert all the methods that reference NotificationChannel and
NotificationChannelGroup to use Object.

Thanks to xian for the suggestion.

Fixes #3098