Fixes 1986
When running tests on Linux platforms, occasionally tests would fail due to an exception like the following:
java.lang.RuntimeException: java.util.concurrent.ExecutionException: java.lang.AssertionError: DB[46] will not dispose from a non-confining thread

The AssertionError is thrown from the sqlite4java library when an attempt was made to close a connection on a different thread to the one it was opened on. All interaction with connections and statements are done by submitting Callables to the dbExecutor ExecutorService, which is a single thread executor.

I believe the issue occurs as there is a race condition in the reset function, which used to look like this:
  public void reset() {
    for (long connectionPtr : connectionsMap.keySet()) {
      close(connectionPtr);
    }
    dbExecutor.shutdown();
    try {
      dbExecutor.awaitTermination(30, TimeUnit.SECONDS);
    } catch (InterruptedException e) {
      throw new RuntimeException(e);
    }
    dbExecutor = Executors.newSingleThreadExecutor();
    connectionsMap.clear();
    statementsMap.clear();
  }

 If a connection was opened between the last call to close(connectionPtr) and this line: dbExecutor = Executors.newSingleThreadExecutor(), then a connection would have been opened on the old thread, and any attempt to close it would be on a new thread.

 The fix is to synchronize all access to anything that interacts with dbExecutor, connectionsMap and statementsMap. A key implication of this is that any work submitted to the dbExecutor must not try to synchroniz on the lock, as it will cause a deadlock.