Support overridden packages with AGP3.0
Allows package name to be overridden in @Config(packageName = ) which is
necessary when R classes are generated in packages other than that
defined in the manifest.

This is still necessary while we parse R classes to obtain the resource
IDs.

When we switch to ARSC files we will no longer need to read R classes
and this can be removed.