Add ShadowInputEventReceiver and no-op consumeBatchedInputEvents.
This method outputs a warning log every time it's called in Robolectric tests, so the message being logged isn't actually indicative of anything other than that a ViewRootImpl traversal occurred. This change just no-ops it to keep the log a bit cleaner and more relevant to actual test failures/output.