Update QueuedWork Resetter for API 26
This is part of an effort to resolve deadlocks that may occur with
SharedPreferences on SDK 26.

In API 26, the implementation of QueuedWork changed from a
ThreadPoolExecutor to a Handler combined with two lists of Runnables. In
this API level, setting the `sHandler` static variable to null led to a
memory leak. Each test that called `QueuedWork.getHandler()` caused a
new Handler to be created and added to `ShadowLooper.loopingLoopers`.

Because the Handler and Runnable lists are being cleared, there's no
need to call `QueuedWork.waitToFinish()` in API 26.