Catch exceptions thrown during filtering and return empty results.
In contrast to the Robolectric ShadowFilter, the actual Android implementation of android.widget.Filter#filter runs asynchronously. One important difference here is that exceptions thrown during performFiltering are caught by the android.widget.Filter.RequestHandler and an empty set of results is returned.

This pull request changes the ShadowFilter to match the Android Filter and also adds a corresponding unit test.