Fix ShadowAccessibilityManager to keep a static instance
1) Shadows.shadowOf(mAccessibilityManager).setEnabled(true)
2) AccessibilityManager.getInstance(mContext).isEnabled()

Calling 1) does not make 2) return true, and 2) is the common pattern
that can be found in the Android framework, so we are not enabling
accessibility in robolectric tests correctly.

The reason is that AccessibilityManager.getInstance(...) calls
ShadowAccessibilityManager.getInstance(...) and it in turn creates new
instance each time, while setEnabled() and other methods only apply to
the previous instance. This can be fixed by keeping a static instance, an
approach similar to the implementation of
AccessibilityManager.getInstance() itself.

https://github.com/robolectric/robolectric/issues/3364