Make build output dir support different working directories
In Android Studio, if the working directory of a test configuration was
not set to the app module directory, the assumed build output directory
could be invalid. This caused a "Resources$NotFoundException" to be
generated.

Update the build output directory lookup to detect this situation. If
the 'intermediates' directory does not exist, attempt to locate the
build output dir by performing a resource lookup on the constants class,
which is always generated in the build output directory.