Some more work to unblock build system api.
Remove ResourceMergerTest - it is  an integration test that assumes
running with the old project.properties parsing code. This won't be the
case for build systems running with the new build system api.

Clean up TestUtil, finding the resources base dir no longer requires
checking a bunch of paths and looking for a TestAndroidManifest.xml