Cleanup of call into PackageParser.parsePackage().
Now hidden behind ShadowPackageManager.callParsePackage().

Instead of using ActivityThread + ResourcesManager to obtain assets and
resources to parse the manifest with we use the actual code from
PackageParser.

We still need to shadow some PackageParser methods because PackageParser
expects AndroidManifest.xml to be at the root of the apk / .ap_ file.
We'd normally be able to just reflectively set
PackageParser.ANDROID_MANIFEST_FILENAME but the compile inlines the
final String unfortuantely. When we switch perminently to binary
resources we can remove all this code :-)