Replace truth.FailureStrategy with truth.FailureMetadata in custom Subjects.
Also changed truth.SubjectFactory to truth.Subject.Factory (plain
renaming) and use method reference instead of anonymous class to create
the factory when applicable.

FailureMetadata, an opaque object to its users, is introduced to replace
FailureStrategy in in custom Subject in order to resolve some existing
flaws of FailureStrategy as well as enable new features to be added to
Truth.

PiperRevId: 175940858