Refine shadow method selection algorithm.
Deprecate @Implements(inheritImplementationMethods). It's inadvisable
because a method meant to shadow a superclass method with the same
signature might be erroneously chosen.

Now, by default, public and protected methods on shadow classes are
considered when searching for methods from the class matching that in
the shadow's @Implements attribute. Shadow class superclasses are
searched for matching methods.