Create a new integration_tests module - CTesque designed to validate Robolectric behaviour against platform behaviour.
Create initial application under test forked from the set of test resources in
the robolectric module. Forking for now so as not to disturb the merging of the
arsc branch. In the near future we will move all Shadow*Test tests to the
testapp module.

Create ResourcesTest based on a stripped down version of
ShadowResourcesTest. Same with AssetManagerTest. These tests will not use any
public Robolectric APIs, rather use the ATSL APIs to allow us to run these
tests both on local host and a device.

PiperOrigin-RevId: 186352510