Eliminate test pollution of static Handler messages
Messages posted to static background Handlers could leak to subsequent tests.
They were stored in the internal 'mMessages' member of MessageQueue. In the
ShadowLooper resetter, schedulers for background Handlers were reset. However,
the message queues themselves were not reset. This could cause subsequent tests
to receive false positives when calling the 'Handler.hasMessages' variants.

This is an issue for the static QueuedWork Handler that is used for
SharedPreferences.

PiperOrigin-RevId: 192786530