Use a unique temp dir to extract and load the SQLite library
When running tests using multiple processes, which is common for large test
suites, there's a race condition caused by each processes' SQLiteLibraryLoader
attempting to extract the native SQLite library concurrently. This is most
noticeable on CI environments, which typically start off with an empty temp
directory. For example, on Linux, each processes' SQLiteLibraryLoader writes
the native SQLite library to the file
`/tmp/robolectric-libs/libsqlite4java.so`. This occasionally causes spurious
test failures.

Update the logic to extract and load the sqlite library from a unique temp
directory. Not only does this eliminate the race, it also removes the need to
check if the native library matches the cached file on disk. From my
benchmarks, there's effectively no performance impact, because the time it
takes to write the embedded native library to disk is about the same as the
time it took to compare the embedded native library to the cached version.

Fixes #3237

PiperOrigin-RevId: 194033430