Add equals() and hashCode() to ShadowDisplay
`ShadowWindowManagerImpl` uses `Display` as a key in a multimap of views. The `Display` class, however, doesn't implement `hashCode()` and `getDefaultDisplay()` returns a new instance of `Display` in each call. As a result, adding views to views always creates a new key, and looking up views doesn't work at all.

This patch adds `equals()` and `hashCode()` to `ShadowDisplay` and fixes #3932.