Stop declaring real files as phony
I've added some logic into Kati to attempt to prevent this type of
thing, since in the majority of cases, we expected a command to only run
if one of its dependencies has changed.

Workaround this specific case by declaring an output that is never
written. DO NOT use this elsewhere without permission. There will
eventually be a check for this workaround as well, but by that time
we'll probably be ready to mark rules that should not be cached.

Test: m RunSettingsLibRoboTests; m RunSettingsLibRoboTests
Test: m Run_robolectric_utils_tests; m Run_robolectric_utils_tests
Change-Id: Ifa9be1b299c0188b9e5f5886a4c80a21de42b0a1