Add ShadowSQLiteOpenHelper, so that we can ignore setIdleConnectionTimeout, which otherwise deadlocks.
Make ShadowStatFs more realistic and convenient, so that we can register statfs
information for a subtree/path prefix.

This is still a bit of a quick-and-dirty solution, in that it doesn't really understand paths, and is algorithmically suboptimal. But neither actually matters. This makes it much more usable than before. In particular, it becomes easy to say "assume no file system is full".

PiperOrigin-RevId: 205334563