1. Add startActivity(Intent) method to ActivityInvoker interface and use it from ActivityScenario instead of directly calling Instrumentation#startActivitySync.
Instrumentation#startActivitySync has an assertion that the resolved Activity's process name equals to the test target package name. This is to ensure that the Activity is launched in the same process as instrumentation but it is possible to run Activities with different process name in a same process if they share the same android:sharedUserId in the both AndroidManifests.

InstrumentationActivityInvoker#startActivity uses Context#startActivity to avoid that assertion. (If the Activity ended up with running in the different process, the test will be timed out.)

LocalActivityInvoker#startActivity continues to use Instrumentation#startActivitySync since its Shadow never check it (and Context#startActivity never start Activity by design).

2. A small bug fix in LocalActivityInvoker#recreate where it doesn't call onRetainNonConfigurationInstance thus losing data over recreation.

PiperOrigin-RevId: 213909555